removed StringEncoder and moved conversions to StringUtlis
